requirements.txt
connexion==2.4.0
Flask==1.1.1
swagger-ui-bundle==0.0.6
PyYAML==5.1.2
requests==2.22.0
pytest==5.3.1
flask-cors==3.0.8
faker==2.0.4